pyproject.toml
[project]
name = "whisper-turbo"
version = "0.1.0"
description = ""
authors = [
    {name = "SomaRe",email = "shivatej82@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper (>=20240930,<20240931)",
    "flask (>=3.1.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
